move `MacOSBrightnessOverrideHandler` to separate file
